GitHub: Try github actions to provide prebuilt sdcard image
This commit is based on PR:
https://github.com/Beckhoff/CX9020/pull/20

We try to keep the GitHub action as close as possibile to the GitLab
based CI. This is no official support for GitHub. If this action breaks,
it has no priority to get fixed. Especially, when issues arise due to
GitHub infrastructure or yaml updates...

Signed-off-by: Steffen Dirkwinkel <s.dirkwinkel@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>